+++ packages/amplify-graphql-api-construct-tests/src/__tests__/deploy-velocity-temporarily-disabled/replace-2-gsis-1k-records.test.ts
@@ -1,10 +1,10 @@
 import {
   COUNT_1_THOUSAND,
-  DURATION_30_MINUTES,
   MUTATION_FOUR_FIELD_CREATE,
   SCHEMA_FOUR_FIELDS_FINAL_TWO_INDEXED,
   SCHEMA_FOUR_FIELDS_INITIAL_TWO_INDEXED,
 } from '../deploy-velocity/deploy-velocity-constants';
+import { DURATION_30_MINUTES } from '../../utils/duration-constants';
 import {
   recordCountDataProvider,
   recordCountDataValidator,