+++ packages/amplify-graphql-api-construct-tests/src/__tests__/group-auth/dynamic-group-auth/test-implementations.ts
@@ -1,3 +1,6 @@
+import { cdkDeploy } from '../../../commands';
+import { addCognitoUserToGroup, createCognitoUser, signInCognitoUser } from '../../../utils';
+import { ONE_MINUTE } from '../../../utils/duration-constants';
 import {
   doCreateRelatedOne,
   doCreateRelatedMany,
@@ -12,6 +15,84 @@ import {
   doGetRelatedMany,
   doListRelatedManies,
 } from '../../graphql-schemas/reference-style-dynamic-group-auth/operation-implementations';
+
+// #region Test setup
+
+export interface CommonSetupInput {
+  projRoot: string;
+  region: string;
+  name: string;
+}
+
+export interface CommonSetupOutput {
+  apiEndpoint: string;
+  group1AccessToken: string;
+  group2AccessToken: string;
+}
+
+export const deployStackAndCreateUsers = async (input: CommonSetupInput): Promise<CommonSetupOutput> => {
+  const { projRoot, region, name } = input;
+  const outputs = await cdkDeploy(projRoot, '--all', { postDeployWaitMs: ONE_MINUTE });
+  const { awsAppsyncApiEndpoint, UserPoolClientId: userPoolClientId, UserPoolId: userPoolId } = outputs[name];
+
+  const apiEndpoint = awsAppsyncApiEndpoint;
+
+  const group1AccessToken = await createTestUser({
+    groupName: 'Group1',
+    region,
+    userPoolId,
+    userPoolClientId,
+  });
+
+  const group2AccessToken = await createTestUser({
+    groupName: 'Group2',
+    region,
+    userPoolId,
+    userPoolClientId,
+  });
+
+  const output: CommonSetupOutput = {
+    apiEndpoint,
+    group1AccessToken,
+    group2AccessToken,
+  };
+
+  return output;
+};
+
+interface CreateUserAndAssignToGroupInput {
+  region: string;
+  userPoolId: string;
+  userPoolClientId: string;
+  groupName: string;
+}
+
+/** Creates a test user and assigns to the specified group */
+const createTestUser = async (input: CreateUserAndAssignToGroupInput): Promise<string> => {
+  const { region, userPoolId, userPoolClientId, groupName } = input;
+  const { username, password } = await createCognitoUser({
+    region,
+    userPoolId,
+  });
+
+  await addCognitoUserToGroup({
+    region,
+    userPoolId,
+    username,
+    group: groupName,
+  });
+
+  const { accessToken } = await signInCognitoUser({
+    username,
+    password,
+    region,
+    userPoolClientId,
+  });
+
+  return accessToken;
+};
+
+// #endregion Test setup
 
 // #region Primary operations
 
@@ -38,6 +119,28 @@ export const testCreatePrimaryDoesNotRedactRelatedForSameOwningGroup = async (
   expect(primary.relatedMany).toBeDefined();
   expect(primary.relatedMany.items.length).toEqual(1);
   expect(primary.relatedMany.items[0].id).toEqual(relatedManyId);
+};
+
+export const testCreatePrimaryRedactsRelatedForSameOwningGroup = async (
+  currentId: number,
+  apiEndpoint: string,
+  accessToken: string,
+): Promise<void> => {
+  const primaryId = `p${currentId}`;
+  const relatedOneId = `ro${currentId}`;
+  const relatedManyId = `rm${currentId}`;
+
+  await doCreateRelatedOne(apiEndpoint, accessToken, relatedOneId, primaryId, ['Group1']);
+  await doCreateRelatedMany(apiEndpoint, accessToken, relatedManyId, primaryId, ['Group1']);
+
+  const result = await doCreatePrimary(apiEndpoint, accessToken, primaryId, ['Group1']);
+  const primary = result.body.data.createPrimary;
+
+  expect(primary).toBeDefined();
+  expect(primary.id).toEqual(primaryId);
+  expect(primary.relatedOne).toBeNull();
+  expect(primary.relatedMany).toBeDefined();
+  expect(primary.relatedMany.items.length).toEqual(0);
 };
 
 export const testCreatePrimaryRedactsRelatedForDifferentOwningGroup = async (
@@ -247,6 +350,29 @@ export const testUpdatePrimaryDoesNotRedactRelatedForSameOwningGroup = async (
   expect(primary.relatedMany.items[0].id).toEqual(relatedManyId);
 };
 
+export const testUpdatePrimaryRedactsRelatedForSameOwningGroup = async (
+  currentId: number,
+  apiEndpoint: string,
+  accessToken: string,
+): Promise<void> => {
+  const primaryId = `p${currentId}`;
+  const relatedOneId = `ro${currentId}`;
+  const relatedManyId = `rm${currentId}`;
+
+  await doCreateRelatedOne(apiEndpoint, accessToken, relatedOneId, primaryId, ['Group1']);
+  await doCreateRelatedMany(apiEndpoint, accessToken, relatedManyId, primaryId, ['Group1']);
+  await doCreatePrimary(apiEndpoint, accessToken, primaryId, ['Group1']);
+
+  const result = await doUpdatePrimary(apiEndpoint, accessToken, primaryId, ['Group1']);
+  const primary = result.body.data.updatePrimary;
+
+  expect(primary).toBeDefined();
+  expect(primary.id).toEqual(primaryId);
+  expect(primary.relatedOne).toBeNull();
+  expect(primary.relatedMany).toBeDefined();
+  expect(primary.relatedMany.items.length).toEqual(0);
+};
+
 export const testUpdatePrimaryRedactsRelatedForDifferentOwningGroup = async (
   currentId: number,
   apiEndpoint: string,
@@ -298,6 +424,26 @@ export const testCreateRelatedOneDoesNotRedactPrimaryForSameOwningGroup = async 
   expect(relatedOne.primary.relatedMany).toBeDefined();
   expect(relatedOne.primary.relatedMany.items.length).toEqual(1);
   expect(relatedOne.primary.relatedMany.items[0].id).toEqual(relatedManyId);
+};
+
+export const testCreateRelatedOneRedactsPrimaryForSameOwningGroup = async (
+  currentId: number,
+  apiEndpoint: string,
+  accessToken: string,
+): Promise<void> => {
+  const primaryId = `p${currentId}`;
+  const relatedOneId = `ro${currentId}`;
+  const relatedManyId = `rm${currentId}`;
+
+  await doCreatePrimary(apiEndpoint, accessToken, primaryId, ['Group1']);
+  await doCreateRelatedMany(apiEndpoint, accessToken, relatedManyId, primaryId, ['Group1']);
+
+  const result = await doCreateRelatedOne(apiEndpoint, accessToken, relatedOneId, primaryId, ['Group1']);
+  const relatedOne = result.body.data.createRelatedOne;
+
+  expect(relatedOne).toBeDefined();
+  expect(relatedOne.id).toEqual(relatedOneId);
+  expect(relatedOne.primary).toBeNull();
 };
 
 export const testCreateRelatedOneRedactsPrimaryForDifferentOwningGroup = async (
@@ -437,6 +583,27 @@ export const testUpdateRelatedOneDoesNotRedactPrimaryForSameOwningGroup = async 
   expect(relatedOne.primary.relatedMany.items[0].id).toEqual(relatedManyId);
 };
 
+export const testUpdateRelatedOneRedactsPrimaryForSameOwningGroup = async (
+  currentId: number,
+  apiEndpoint: string,
+  accessToken: string,
+): Promise<void> => {
+  const primaryId = `p${currentId}`;
+  const relatedOneId = `ro${currentId}`;
+  const relatedManyId = `rm${currentId}`;
+
+  await doCreatePrimary(apiEndpoint, accessToken, primaryId, ['Group1']);
+  await doCreateRelatedMany(apiEndpoint, accessToken, relatedManyId, primaryId, ['Group1']);
+  await doCreateRelatedOne(apiEndpoint, accessToken, relatedOneId, primaryId, ['Group1']);
+
+  const result = await doUpdateRelatedOne(apiEndpoint, accessToken, relatedOneId, primaryId, ['Group1']);
+  const relatedOne = result.body.data.updateRelatedOne;
+
+  expect(relatedOne).toBeDefined();
+  expect(relatedOne.id).toEqual(relatedOneId);
+  expect(relatedOne.primary).toBeNull();
+};
+
 export const testUpdateRelatedOneRedactsPrimaryForDifferentOwningGroup = async (
   currentId: number,
   apiEndpoint: string,
@@ -483,6 +650,26 @@ export const testCreateRelatedManyDoesNotRedactPrimaryForSameOwningGroup = async
 
   expect(relatedMany.primary.relatedOne).toBeDefined();
   expect(relatedMany.primary.relatedOne.id).toEqual(relatedOneId);
+};
+
+export const testCreateRelatedManyRedactsPrimaryForSameOwningGroup = async (
+  currentId: number,
+  apiEndpoint: string,
+  accessToken: string,
+): Promise<void> => {
+  const primaryId = `p${currentId}`;
+  const relatedOneId = `ro${currentId}`;
+  const relatedManyId = `rm${currentId}`;
+
+  await doCreatePrimary(apiEndpoint, accessToken, primaryId, ['Group1']);
+  await doCreateRelatedOne(apiEndpoint, accessToken, relatedOneId, primaryId, ['Group1']);
+
+  const result = await doCreateRelatedMany(apiEndpoint, accessToken, relatedManyId, primaryId, ['Group1']);
+  const relatedMany = result.body.data.createRelatedMany;
+
+  expect(relatedMany).toBeDefined();
+  expect(relatedMany.id).toEqual(relatedManyId);
+  expect(relatedMany.primary).toBeNull();
 };
 
 export const testCreateRelatedManyRedactsPrimaryForDifferentOwningGroup = async (
@@ -619,6 +806,27 @@ export const testUpdateRelatedManyDoesNotRedactPrimaryForSameOwningGroup = async
 
   expect(relatedMany.primary.relatedOne).toBeDefined();
   expect(relatedMany.primary.relatedOne.id).toEqual(relatedOneId);
+};
+
+export const testUpdateRelatedManyRedactsPrimaryForSameOwningGroup = async (
+  currentId: number,
+  apiEndpoint: string,
+  accessToken: string,
+): Promise<void> => {
+  const primaryId = `p${currentId}`;
+  const relatedOneId = `ro${currentId}`;
+  const relatedManyId = `rm${currentId}`;
+
+  await doCreatePrimary(apiEndpoint, accessToken, primaryId, ['Group1']);
+  await doCreateRelatedOne(apiEndpoint, accessToken, relatedOneId, primaryId, ['Group1']);
+  await doCreateRelatedMany(apiEndpoint, accessToken, relatedManyId, primaryId, ['Group1']);
+
+  const result = await doUpdateRelatedMany(apiEndpoint, accessToken, relatedManyId, primaryId, ['Group1']);
+  const relatedMany = result.body.data.updateRelatedMany;
+
+  expect(relatedMany).toBeDefined();
+  expect(relatedMany.id).toEqual(relatedManyId);
+  expect(relatedMany.primary).toBeNull();
 };
 
 export const testUpdateRelatedManyRedactsPrimaryForDifferentOwningGroup = async (