+++ packages/amplify-graphql-api-construct-tests/src/__tests__/owner-auth/subscriptions-off/assoc-field-subscriptions-off.test.ts
@@ -10,10 +10,11 @@ import {
   dbDetailsToModelDataSourceStrategy,
   signInCognitoUser,
   TestDefinition,
-  writeStackPrefix,
+  writeStackConfig,
   writeTestDefinitions,
 } from '../../../utils';
 import { SqlDatabaseDetails, SqlDatatabaseController } from '../../../sql-datatabase-controller';
+import { DURATION_1_HOUR, ONE_MINUTE } from '../../../utils/duration-constants';
 import {
   testCreatePrimaryRedactedForDifferentOwners,
   testCreatePrimaryVisibleForSameOwner,
@@ -41,15 +42,15 @@ import {
   testUpdateRelatedManyVisibleForSameOwner,
   testUpdateRelatedOneRedactedForDifferentOwners,
   testUpdateRelatedOneVisibleForSameOwner,
-} from './test-implementations';
+} from '../assoc-field/test-implementations';
 
-jest.setTimeout(1000 * 60 * 60 /* 1 hour */);
+jest.setTimeout(DURATION_1_HOUR);
 
 // Each of these tests asserts that restricted fields in associated types are properly redacted. To assert this, we create the relationship
 // records in an order so that the type we're asserting on comes LAST. By "prepopulating" the associated records before creating the source
 // record, we ensure that the selection set is fully populated with relationship data, and can therefore assert that restricted fields on
 // the associated records are redacted.
-describe('Associated fields protected by owner auth control visibility appropriately', () => {
+describe('Associated fields protected by owner auth control visibility appropriately subscriptions off', () => {
   const region = process.env.CLI_REGION ?? 'us-west-2';
   const baseProjFolderName = path.basename(__filename, '.test.ts');
 
@@ -111,12 +112,12 @@ describe('Associated fields protected by owner auth control visibility appropria
       const name = await initCDKProject(projRoot, templatePath);
 
       const primarySchemaPath = path.resolve(
-        path.join(__dirname, '..', '..', 'graphql-schemas', 'reference-style-owner-auth', 'schema-primary.graphql'),
+        path.join(__dirname, '..', '..', 'graphql-schemas', 'reference-style-owner-auth', 'schema-primary-subscriptions-off.graphql'),
       );
       const primarySchema = fs.readFileSync(primarySchemaPath).toString();
 
       const relatedSchemaPath = path.resolve(
-        path.join(__dirname, '..', '..', 'graphql-schemas', 'reference-style-owner-auth', 'schema-related.graphql'),
+        path.join(__dirname, '..', '..', 'graphql-schemas', 'reference-style-owner-auth', 'schema-related-subscriptions-off.graphql'),
       );
       const relatedSchema = fs.readFileSync(relatedSchemaPath).toString();
 
@@ -127,7 +128,7 @@ describe('Associated fields protected by owner auth control visibility appropria
         },
       };
 
-      writeStackPrefix('AFDdbDdb', projRoot);
+      writeStackConfig(projRoot, { prefix: 'AFDdbDdb' });
       writeTestDefinitions(testDefinitions, projRoot);
 
       const outputs = await cdkDeploy(projRoot, '--all');
@@ -305,12 +306,12 @@ describe('Associated fields protected by owner auth control visibility appropria
       const name = await initCDKProject(projRoot, templatePath);
 
       const primarySchemaPath = path.resolve(
-        path.join(__dirname, '..', '..', 'graphql-schemas', 'reference-style-owner-auth', 'schema-primary.graphql'),
+        path.join(__dirname, '..', '..', 'graphql-schemas', 'reference-style-owner-auth', 'schema-primary-subscriptions-off.graphql'),
       );
       const primarySchema = fs.readFileSync(primarySchemaPath).toString();
 
       const relatedSchemaPath = path.resolve(
-        path.join(__dirname, '..', '..', 'graphql-schemas', 'reference-style-owner-auth', 'schema-related.graphql'),
+        path.join(__dirname, '..', '..', 'graphql-schemas', 'reference-style-owner-auth', 'schema-related-subscriptions-off.graphql'),
       );
       const relatedSchema = fs.readFileSync(relatedSchemaPath).toString();
 
@@ -321,10 +322,10 @@ describe('Associated fields protected by owner auth control visibility appropria
         },
       };
 
-      writeStackPrefix('AFSqlSql', projRoot);
+      writeStackConfig(projRoot, { prefix: 'AFSqlSql' });
       writeTestDefinitions(testDefinitions, projRoot);
 
-      const outputs = await cdkDeploy(projRoot, '--all');
+      const outputs = await cdkDeploy(projRoot, '--all', { postDeployWaitMs: ONE_MINUTE });
       const { awsAppsyncApiEndpoint, UserPoolClientId: userPoolClientId, UserPoolId: userPoolId } = outputs[name];
 
       apiEndpoint = awsAppsyncApiEndpoint;
@@ -499,12 +500,12 @@ describe('Associated fields protected by owner auth control visibility appropria
       const name = await initCDKProject(projRoot, templatePath);
 
       const primarySchemaPath = path.resolve(
-        path.join(__dirname, '..', '..', 'graphql-schemas', 'reference-style-owner-auth', 'schema-primary.graphql'),
+        path.join(__dirname, '..', '..', 'graphql-schemas', 'reference-style-owner-auth', 'schema-primary-subscriptions-off.graphql'),
       );
       const primarySchema = fs.readFileSync(primarySchemaPath).toString();
 
       const relatedSchemaPath = path.resolve(
-        path.join(__dirname, '..', '..', 'graphql-schemas', 'reference-style-owner-auth', 'schema-related.graphql'),
+        path.join(__dirname, '..', '..', 'graphql-schemas', 'reference-style-owner-auth', 'schema-related-subscriptions-off.graphql'),
       );
       const relatedSchema = fs.readFileSync(relatedSchemaPath).toString();
 
@@ -519,10 +520,10 @@ describe('Associated fields protected by owner auth control visibility appropria
         },
       };
 
-      writeStackPrefix('AFSqlDdb', projRoot);
+      writeStackConfig(projRoot, { prefix: 'AFSqlDdb' });
       writeTestDefinitions(testDefinitions, projRoot);
 
-      const outputs = await cdkDeploy(projRoot, '--all');
+      const outputs = await cdkDeploy(projRoot, '--all', { postDeployWaitMs: ONE_MINUTE });
       const { awsAppsyncApiEndpoint, UserPoolClientId: userPoolClientId, UserPoolId: userPoolId } = outputs[name];
 
       apiEndpoint = awsAppsyncApiEndpoint;
@@ -697,13 +698,13 @@ describe('Associated fields protected by owner auth control visibility appropria
       const name = await initCDKProject(projRoot, templatePath);
 
       const primarySchemaPath = path.resolve(
-        path.join(__dirname, '..', '..', 'graphql-schemas', 'reference-style-owner-auth', 'schema-primary.graphql'),
+        path.join(__dirname, '..', '..', 'graphql-schemas', 'reference-style-owner-auth', 'schema-primary-subscriptions-off.graphql'),
       );
 
       const primarySchema = fs.readFileSync(primarySchemaPath).toString();
 
       const relatedSchemaPath = path.resolve(
-        path.join(__dirname, '..', '..', 'graphql-schemas', 'reference-style-owner-auth', 'schema-related.graphql'),
+        path.join(__dirname, '..', '..', 'graphql-schemas', 'reference-style-owner-auth', 'schema-related-subscriptions-off.graphql'),
       );
       const relatedSchema = fs.readFileSync(relatedSchemaPath).toString();
 
@@ -718,10 +719,10 @@ describe('Associated fields protected by owner auth control visibility appropria
         },
       };
 
-      writeStackPrefix('AFDdbSql', projRoot);
+      writeStackConfig(projRoot, { prefix: 'AFDdbSql' });
       writeTestDefinitions(testDefinitions, projRoot);
 
-      const outputs = await cdkDeploy(projRoot, '--all');
+      const outputs = await cdkDeploy(projRoot, '--all', { postDeployWaitMs: ONE_MINUTE });
       const { awsAppsyncApiEndpoint, UserPoolClientId: userPoolClientId, UserPoolId: userPoolId } = outputs[name];
 
       apiEndpoint = awsAppsyncApiEndpoint;