+++ packages/amplify-graphql-api-construct-tests/src/__tests__/relationships/references/references-ddbprimary-sqlrelated.test.ts
@@ -1,12 +1,13 @@
 /* eslint-disable import/namespace */
-import { DDB_AMPLIFY_MANAGED_DATASOURCE_STRATEGY } from '@aws-amplify/graphql-transformer-core';
-import { createNewProjectDir, deleteProjectDir } from 'amplify-category-api-e2e-core';
+import * as path from 'path';
 import * as fs from 'fs-extra';
 import * as generator from 'generate-password';
-import * as path from 'path';
+import { DDB_AMPLIFY_MANAGED_DATASOURCE_STRATEGY } from '@aws-amplify/graphql-transformer-core';
+import { createNewProjectDir, deleteProjectDir } from 'amplify-category-api-e2e-core';
 import { cdkDeploy, cdkDestroy, initCDKProject } from '../../../commands';
 import { SqlDatabaseDetails, SqlDatatabaseController } from '../../../sql-datatabase-controller';
-import { TestDefinition, dbDetailsToModelDataSourceStrategy, writeStackPrefix, writeTestDefinitions } from '../../../utils';
+import { TestDefinition, dbDetailsToModelDataSourceStrategy, writeStackConfig, writeTestDefinitions } from '../../../utils';
+import { DURATION_1_HOUR, ONE_MINUTE } from '../../../utils/duration-constants';
 import {
   testPrimaryContainsAssociated,
   testPrimaryCpkSkOneContainsAssociated,
@@ -20,7 +21,7 @@ import {
   testRelatedOneCpkSkTwoContainsAssociated,
 } from './test-implementations';
 
-jest.setTimeout(1000 * 60 * 60 /* 1 hour */);
+jest.setTimeout(DURATION_1_HOUR);
 
 describe('References relationships', () => {
   const region = process.env.CLI_REGION ?? 'us-west-2';
@@ -95,7 +96,7 @@ describe('References relationships', () => {
 
     beforeAll(async () => {
       projRoot = await createNewProjectDir(projFolderName);
-      const templatePath = path.resolve(path.join(__dirname, '..', '..', 'backends', 'configurable-sandbox-stack'));
+      const templatePath = path.resolve(path.join(__dirname, '..', '..', 'backends', 'configurable-stack'));
       const name = await initCDKProject(projRoot, templatePath);
 
       const primarySchemaPath = path.resolve(path.join(__dirname, 'graphql', 'schema-primary.graphql'));
@@ -127,10 +128,10 @@ describe('References relationships', () => {
         },
       };
 
-      writeStackPrefix('RefSqlDdb', projRoot);
+      writeStackConfig(projRoot, { prefix: 'RefSqlDdb', useSandbox: true });
       writeTestDefinitions(testDefinitions, projRoot);
 
-      const outputs = await cdkDeploy(projRoot, '--all');
+      const outputs = await cdkDeploy(projRoot, '--all', { postDeployWaitMs: ONE_MINUTE });
       apiEndpoint = outputs[name].awsAppsyncApiEndpoint;
       apiKey = outputs[name].awsAppsyncApiKey;
     });