+++ scripts/split-e2e-tests.ts
@@ -6,14 +6,23 @@ import * as yaml from 'js-yaml';
 type TestRegion = {
   name: string;
   optIn: boolean;
+  cognitoSupported: boolean;
+  betaLayerDeployed: boolean; // is the beta layer deployed in this region
 };
 
 const REPO_ROOT = join(__dirname, '..');
 
 const supportedRegionsPath = join(REPO_ROOT, 'scripts', 'e2e-test-regions.json');
-const suportedRegions: TestRegion[] = JSON.parse(fs.readFileSync(supportedRegionsPath, 'utf-8'));
-const testRegions = suportedRegions.map((region) => region.name);
-const nonOptInRegions = suportedRegions.filter((region) => !region.optIn).map((region) => region.name);
+const supportedRegions: TestRegion[] = JSON.parse(fs.readFileSync(supportedRegionsPath, 'utf-8'));
+const testRegions = supportedRegions.map((region) => region.name);
+const supportedRegionsByRegionName: Record<string, TestRegion> = supportedRegions.reduce(
+  (acc, region) => ({ ...acc, [region.name]: region }),
+  {},
+);
+
+// list of regions the beta layer is not deployed in
+// the tests should not use these regions when using the beta layer
+const BETA_LAYER_NOT_DEPLOYED = supportedRegions.filter((region) => !region.betaLayerDeployed).map((region) => region.name);
 
 // https://github.com/aws-amplify/amplify-cli/blob/d55917fd83140817a4447b3def1736f75142df44/packages/amplify-provider-awscloudformation/src/aws-regions.js#L4-L17
 const v1TransformerSupportedRegionsPath = join(REPO_ROOT, 'scripts', 'v1-transformer-supported-regions.json');
@@ -83,6 +92,7 @@ const TEST_TIMINGS_PATH = join(REPO_ROOT, 'scripts', 'test-timings.data.json');
 const CODEBUILD_CONFIG_BASE_PATH = join(REPO_ROOT, 'codebuild_specs', 'e2e_workflow_base.yml');
 const CODEBUILD_GENERATE_CONFIG_PATH = join(REPO_ROOT, 'codebuild_specs', 'e2e_workflow.yml');
 const CODEBUILD_DEBUG_CONFIG_PATH = join(REPO_ROOT, 'codebuild_specs', 'debug_workflow.yml');
+
 const RUN_SOLO: (string | RegExp)[] = [
   'src/__tests__/apigw.test.ts',
   'src/__tests__/api_2.test.ts',
@@ -167,6 +177,13 @@ const RUN_IN_NON_OPT_IN_REGIONS: (string | RegExp)[] = [
   'src/__tests__/ddb-iam-access.test.ts',
 ];
 
+const RUN_IN_COGNITO_REGIONS: (string | RegExp)[] = [
+  /src\/__tests__\/.*userpool.*\.test\.ts/,
+  /src\/__tests__\/group-auth\/.*\.test\.ts/,
+  /src\/__tests__\/owner-auth\/.*\.test\.ts/,
+  /src\/__tests__\/restricted-field-auth\/.*\.test\.ts/,
+];
+
 const RUN_IN_V1_TRANSFORMER_REGIONS = ['src/__tests__/schema-searchable.test.ts'];
 
 const DEBUG_FLAG = '--debug';
@@ -204,7 +221,12 @@ const getTestNameFromPath = (testSuitePath: string, region?: string): string => 
   return testSuitePath.substring(startIndex, endIndex).split('.e2e').join('').split('.').join('-').concat(regionSuffix);
 };
 
-const splitTests = (baseJobLinux: any, testDirectory: string, pickTests?: (testSuites: string[]) => string[]): BatchBuildJob[] => {
+const splitTests = (
+  baseJobLinux: any,
+  testDirectory: string,
+  useBetaLayer: boolean = false,
+  pickTests?: (testSuites: string[]) => string[],
+): BatchBuildJob[] => {
   const output: any[] = [];
   let testSuites = getTestFiles(testDirectory);
   if (pickTests && typeof pickTests === 'function') {
@@ -220,6 +242,7 @@ const splitTests = (baseJobLinux: any, testDirectory: string, pickTests?: (testS
     const runtimeB = testFileRunTimes.find((t: any) => t.test === b)?.medianRuntime ?? 30;
     return runtimeA - runtimeB;
   });
+
   const generateJobsForOS = (os: OSType): CandidateJob[] => {
     const soloJobs = [];
     let jobIdx = 0;
@@ -231,12 +254,10 @@ const splitTests = (baseJobLinux: any, testDirectory: string, pickTests?: (testS
       const USE_PARENT = USE_PARENT_ACCOUNT.some((usesParent) => test.startsWith(usesParent));
 
       if (RUN_SOLO.find((solo) => test === solo || test.match(solo))) {
-        if (RUN_IN_ALL_REGIONS.find((allRegions) => test === allRegions || test.match(allRegions))) {
-          const shouldRunInNonOptInRegion = RUN_IN_NON_OPT_IN_REGIONS.find(
-            (nonOptInTest) => test.toLowerCase() === nonOptInTest || test.toLowerCase().match(nonOptInTest),
-          );
-          const regionsToRunTest = shouldRunInNonOptInRegion ? nonOptInRegions : testRegions;
-          regionsToRunTest.forEach((region) => {
+        if (RUN_IN_ALL_REGIONS.find((allRegionsTest) => test === allRegionsTest || test.match(allRegionsTest))) {
+          // always run these jobs in regions that do not have the beta layer deployed
+          const candidateRegions = filterCandidateRegions(test, testRegions, false);
+          candidateRegions.forEach((region) => {
             const newSoloJob = createJob(os, jobIdx, true);
             jobIdx++;
             newSoloJob.tests.push(test);
@@ -252,14 +273,14 @@ const splitTests = (baseJobLinux: any, testDirectory: string, pickTests?: (testS
         if (USE_PARENT) {
           newSoloJob.useParentAccount = true;
         }
-        setJobRegion(test, newSoloJob, jobIdx);
+        setJobRegion(test, newSoloJob, jobIdx, useBetaLayer);
         soloJobs.push(newSoloJob);
         continue;
       }
 
       // add the test
       currentJob.tests.push(test);
-      setJobRegion(test, currentJob, jobIdx);
+      setJobRegion(test, currentJob, jobIdx, useBetaLayer);
       if (USE_PARENT) {
         currentJob.useParentAccount = true;
       }
@@ -272,6 +293,7 @@ const splitTests = (baseJobLinux: any, testDirectory: string, pickTests?: (testS
     }
     return [...osJobs, ...soloJobs];
   };
+
   const linuxJobs = generateJobsForOS('l');
   const getIdentifier = (names: string): string => `${names.replace(/-/g, '_')}`.substring(0, 127);
   const result: any[] = [];
@@ -297,7 +319,7 @@ const splitTests = (baseJobLinux: any, testDirectory: string, pickTests?: (testS
   return result;
 };
 
-const setJobRegion = (test: string, job: CandidateJob, jobIdx: number) => {
+const setJobRegion = (test: string, job: CandidateJob, jobIdx: number, useBetaLayer: boolean): void => {
   const FORCE_REGION = Object.keys(FORCE_REGION_MAP).find((key) => {
     const testName = getTestNameFromPath(test);
     return testName.startsWith(key);
@@ -314,20 +336,48 @@ const setJobRegion = (test: string, job: CandidateJob, jobIdx: number) => {
     return;
   }
 
-  // Parent E2E account does not have opt-in regions. Choose non-opt-in region.
-  // If the tests are explicitly specified as to be run in non-opt-in regions, follow that.
-  if (
-    RUN_IN_NON_OPT_IN_REGIONS.find((nonOptInTest) => test.toLowerCase() === nonOptInTest || test.toLowerCase().match(nonOptInTest)) ||
-    USE_PARENT_ACCOUNT.some((usesParent) => test.startsWith(usesParent))
-  ) {
-    if (!nonOptInRegions.includes(job.region)) {
-      job.region = nonOptInRegions[jobIdx % nonOptInRegions.length];
-    }
+  const candidateRegions = filterCandidateRegions(test, testRegions, useBetaLayer);
+
+  if (candidateRegions.length === 0) {
+    throw new Error(`No candidate regions found for test ${test}`);
   }
+
+  job.region = candidateRegions[jobIdx % candidateRegions.length];
+};
+
+const filterCandidateRegions = (test: string, candidateRegions: string[], useBetaLayer: boolean): string[] => {
+  let resolvedRegions = [...candidateRegions];
+
+  // Parent E2E account does not have opt-in regions. Choose non-opt-in region.
+  const shouldUseParentAccount = USE_PARENT_ACCOUNT.some((usesParent) => test.startsWith(usesParent));
+
+  // If the tests are explicitly specified as to be run in opt-in regions, respect that.
+  const shouldRunInNonOptInRegion = RUN_IN_NON_OPT_IN_REGIONS.some(
+    (nonOptInTest) => test.toLowerCase() === nonOptInTest || test.toLowerCase().match(nonOptInTest),
+  );
+
+  if (shouldUseParentAccount || shouldRunInNonOptInRegion) {
+    resolvedRegions = resolvedRegions.filter((region) => !supportedRegionsByRegionName[region].optIn);
+  }
+
+  // Some tests require Cognito User Pools or Identity Pools
+  const shouldRunInCognitoRegion = RUN_IN_COGNITO_REGIONS.some(
+    (cognitoTest) => test.toLowerCase() === cognitoTest || test.toLowerCase().match(cognitoTest),
+  );
+  if (shouldRunInCognitoRegion) {
+    resolvedRegions = resolvedRegions.filter((region) => supportedRegionsByRegionName[region].cognitoSupported);
+  }
+
+  if (useBetaLayer) {
+    resolvedRegions = resolvedRegions.filter((region) => !BETA_LAYER_NOT_DEPLOYED.includes(region));
+  }
+
+  return resolvedRegions;
 };
 
 const main = (): void => {
-  const filteredTests = process.argv.slice(2);
+  const useBetaLayer = process.argv[2] === 'beta';
+  const filteredTests = process.argv.slice(3);
   const configBase: ConfigBase = loadConfigBase();
   const baseBuildGraph = configBase.batch['build-graph'];
 
@@ -342,6 +392,7 @@ const main = (): void => {
         'depend-on': ['publish_to_local_registry'],
       },
       join(REPO_ROOT, 'packages', 'amplify-e2e-tests'),
+      useBetaLayer,
     ),
     ...splitTests(
       {
@@ -353,6 +404,7 @@ const main = (): void => {
         'depend-on': ['publish_to_local_registry'],
       },
       join(REPO_ROOT, 'packages', 'amplify-graphql-api-construct-tests'),
+      useBetaLayer,
     ),
     ...splitTests(
       {
@@ -364,6 +416,7 @@ const main = (): void => {
         'depend-on': ['publish_to_local_registry'],
       },
       join(REPO_ROOT, 'packages', 'graphql-transformers-e2e-tests'),
+      useBetaLayer,
     ),
   ];