+++ packages/amplify-e2e-tests/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.22.4](https://github.com/aws-amplify/amplify-category-api/compare/amplify-category-api-e2e-tests@3.22.3...amplify-category-api-e2e-tests@3.22.4) (2024-06-25)
+
+**Note:** Version bump only for package amplify-category-api-e2e-tests
+
+## [3.22.3](https://github.com/aws-amplify/amplify-category-api/compare/amplify-category-api-e2e-tests@3.22.2...amplify-category-api-e2e-tests@3.22.3) (2024-05-15)
+
+**Note:** Version bump only for package amplify-category-api-e2e-tests
+
+## [3.22.2](https://github.com/aws-amplify/amplify-category-api/compare/amplify-category-api-e2e-tests@3.22.1...amplify-category-api-e2e-tests@3.22.2) (2024-05-10)
+
+**Note:** Version bump only for package amplify-category-api-e2e-tests
+
 ## [3.22.1](https://github.com/aws-amplify/amplify-category-api/compare/amplify-category-api-e2e-tests@3.22.0...amplify-category-api-e2e-tests@3.22.1) (2024-05-01)
 
 **Note:** Version bump only for package amplify-category-api-e2e-tests