+++ packages/amplify-graphql-api-construct/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.10.0](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-api-construct@1.9.5...@aws-amplify/graphql-api-construct@1.10.0) (2024-06-25)
+
+### Features
+
+- allow subscriptions to inherit primary model auth rules for relational fields behind a feature flag ([#2649](https://github.com/aws-amplify/amplify-category-api/issues/2649)) ([56a853a](https://github.com/aws-amplify/amplify-category-api/commit/56a853ace0026de97395cfa17ca156cf360ac5c2))
+
+## [1.9.5](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-api-construct@1.9.4...@aws-amplify/graphql-api-construct@1.9.5) (2024-06-06)
+
+**Note:** Version bump only for package @aws-amplify/graphql-api-construct
+
+## [1.9.4](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-api-construct@1.9.3...@aws-amplify/graphql-api-construct@1.9.4) (2024-06-04)
+
+**Note:** Version bump only for package @aws-amplify/graphql-api-construct
+
+## [1.9.3](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-api-construct@1.9.2...@aws-amplify/graphql-api-construct@1.9.3) (2024-05-15)
+
+**Note:** Version bump only for package @aws-amplify/graphql-api-construct
+
+## [1.9.2](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-api-construct@1.9.1...@aws-amplify/graphql-api-construct@1.9.2) (2024-05-10)
+
+**Note:** Version bump only for package @aws-amplify/graphql-api-construct
+
 ## [1.9.1](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-api-construct@1.9.0...@aws-amplify/graphql-api-construct@1.9.1) (2024-05-01)
 
 **Note:** Version bump only for package @aws-amplify/graphql-api-construct